Run differential association tests between between all combinations of a factor variable with DESeq2. Can be used as post-hoc test for regression.

Run differential association tests between between all combinations of a factor variable with DESeq2. Can be used as post-hoc test for regression.

combinatorial_association(
  ps,
  variable,
  tax = "genus",
  confounders = NULL,
  min_abundance = 10,
  in_samples = 0.1,
  independent_weighting = TRUE,
  standardize = TRUE,
  shrink = TRUE
)

Arguments

ps

A phyloseq object containing the taxa counts.

variable

The factor variable to permute.

tax

The taxa level on which to run differential tests. Defaults to genus.

confounders

A character vector containing the confounders that should be used.

min_abundance

Minimum required number of average counts for a taxa.

in_samples

Taxa must be present in at least this fraction of samples.

independent_weighting

Whether to adjust p values by independent weighting or normal Benjamini-Hochberg. factors.

standardize

Whether to standardize continuous variables to a mean of zero and a variance of 1. If True log fold changes for those variables denote are relative to a change of one standard deviation in the variable value.

shrink

Whether to return shrunken log fold changes. Defaults to true.

Value

A data.table containing the results.

Examples

 NULL
#> NULL